ecr-repository/requirements.txt
amqp==2.6.1
attrs==20.3.0
boto3==1.17.50
botocore==1.20.50
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
dnspython==1.16.0
eventlet==0.30.2
greenlet==1.0.0
h5py==3.2.1
idna==2.10
imageio==2.9.0
imantics==0.1.12
imgaug==0.4.0
jmespath==0.10.0
jsonschema==3.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==4.6.11
labelme2coco==0.1.2
lxml==4.6.3
matplotlib==3.4.1
mock==4.0.3
nameko==2.13.0
networkx==2.5.1
numpy==1.20.2
opencv-python==4.5.1.48
path==15.1.2
path.py==12.5.0
Pillow==8.2.0
pixellib==0.6.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.0
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.6
scikit-image==0.18.1
scipy==1.6.2
Shapely==1.7.1
six==1.15.0
tifffile==2021.4.8
urllib3==1.26.4
vine==1.3.0
Werkzeug==1.0.1
wrapt==1.12.1
xmljson==0.2.1
